@@ -1,9 +1,9 @@
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Http.Features;
-using OpenDataBackend.Exceptions;
+using OpenData.Exceptions;
  
-namespace OpenDataBackend.Middleware
+namespace OpenData.Middleware
 {
     public class HttpExceptionMiddleware
     {
